+++ R/get_package.R
@@ -23,7 +23,7 @@
 #'   these did not detect the selected animals.
 #' - `receivers`: Acoustic receivers for the selected deployments, as returned
 #'   by [get_acoustic_receivers()].
-#' - `references`: References for ETN, the R package, the animal project and
+#' - `Bibliography`: A bibliography for ETN, the R package, the animal project and
 #'   acoustic projects that returned detections.
 #'   It is recommended to cite these when using the dataset.
 #'
@@ -135,29 +135,9 @@ get_package <- function(animal_project_code) {
     receiver_id = receiver_ids
   )
 
-  ## References ----
-  cli::cli_li("Getting {.val references}.")
-  etn_ref <- paste(
-    "European Tracking Network - Data Platform.",
-    "Flanders Marine Institute (VLIZ)"
-  )
-  animal_ref <-
-    get_animal_projects(
-      animal_project_code = animal_project_code,
-      citation = TRUE
-    ) |>
-    dplyr::pull("citation")
-  acoustic_refs <-
-    get_acoustic_projects(
-      acoustic_project_code = acoustic_project_codes,
-      citation = TRUE
-    ) |>
-    dplyr::pull("citation")
-  references <-
-    dplyr::tibble(
-      reference_for = c("ETN", animal_project_code, acoustic_project_codes),
-      reference = c(etn_ref, animal_ref, acoustic_refs)
-    )
+  ## Bibliography ----
+  cli::cli_li("Getting {.val bibliography}.")
+  bibliography <- get_bibliography(detections)
 
   cli::cli_end()
 
@@ -180,7 +160,7 @@ get_package <- function(animal_project_code) {
     add_resource("detections", detections) |>
     add_resource("deployments", deployments) |>
     add_resource("receivers", receivers) |>
-    add_resource("references", references) |>
+    add_resource("bibliography", bibliography) |>
     append(c(
       id = doi,
       name = tolower(animal_project_code)